+++ modules/siddhi-launcher/samples/sample-clients/tcp-client/src/main/java/org/wso2/sp/tcp/client/StoreTestClient.java
@@ -21,9 +21,9 @@ package org.wso2.sp.tcp.client;
 import org.apache.log4j.Logger;
 import org.wso2.extension.siddhi.io.tcp.transport.TCPNettyClient;
 import org.wso2.extension.siddhi.map.binary.sinkmapper.BinaryEventConverter;
-import org.wso2.siddhi.core.event.Event;
-import org.wso2.siddhi.core.exception.ConnectionUnavailableException;
-import org.wso2.siddhi.query.api.definition.Attribute;
+import io.siddhi.core.event.Event;
+import io.siddhi.core.exception.ConnectionUnavailableException;
+import io.siddhi.query.api.definition.Attribute;
 
 import java.io.IOException;
 import java.util.ArrayList;
